+++ lib/ui/blocs/post_details/post_details_state.dart
@@ -42,10 +42,9 @@ class PostDetailsLoaded extends PostDetailsState {
   int get commentsCount => comments.length;
 
   /// Returns the list of reactions associated to the post.
-  List<Reaction> get reactions =>
-      post.reactions.where((element) => !element.isLike).toList();
+  List<Reaction> get reactions => post.reactions;
 
-  /// Returns the number of reactions (excluding the likes) that the
+  /// Returns the number of reactions that the
   /// current post has.
   int get reactionsCount => reactions.length;