+++ lib/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/post_comments_list/widgets/post_comment_item/index.dart
@@ -54,8 +54,6 @@ class PostCommentItem extends StatelessWidget {
       children: <Widget>[
         PostCommentAction(size: ICON_SIZE, post: comment),
         const SizedBox(width: ICON_SIZE),
-        PostLikeAction(size: ICON_SIZE, isLiked: isLiked, post: comment),
-        const SizedBox(width: ICON_SIZE),
         PostAddReactionAction(size: ICON_SIZE, post: comment),
       ],
     );