+++ lib/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/post_details_reactions_list/widgets/empty_reactions/index.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:mooncake/ui/ui.dart';
+
+class EmptyReactions extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        mainAxisAlignment: MainAxisAlignment.spaceAround,
+        children: <Widget>[
+          Expanded(child: Center()),
+          Expanded(
+            child: Center(
+              child: Image.asset("assets/images/smile.png"),
+            ),
+          ),
+          Expanded(
+            child: Center(
+              child: Text(
+                PostsLocalizations.of(context)
+                    .translate(Messages.noReactionsYet),
+                textAlign: TextAlign.center,
+              ),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}