+++ lib/ui/screens/post_details_screen/widgets/post_details_main_content/widgets/post_details_reactions_list/widgets/export.dart
@@ -1,1 +1,2 @@
-export 'post_details_reactions_list/index.dart';
+export 'empty_reactions/index.dart';
+export 'reactions_list/index.dart';