+++ test/ui/widgets/post_content/widgets/post_poll_content/index_test.dart
@@ -36,7 +36,7 @@ void main() {
 
     MockAccountBloc mockAccountBloc = MockAccountBloc();
     when(mockAccountBloc.state).thenReturn(LoggedIn.initial(userAccount));
-    when(mockAccountBloc.state).thenReturn(LoggedIn.initial(userAccount));
+
     await tester.pumpWidget(
       makeTestableWidget(
         child: MultiBlocProvider(